+++ R/pin_exists.R
@@ -7,10 +7,10 @@ pin_exists <- function(board, name, ...) {
   UseMethod("pin_exists")
 }
 
-check_pin_exists <- function(board, name) {
+check_pin_exists <- function(board, name, call = caller_env()) {
   if (pin_exists(board, name)) {
     invisible()
   } else {
-    abort_pin_missing(name)
+    abort_pin_missing(name, call = call)
   }
 }