+++ R/pins-package.R
@@ -5,6 +5,7 @@
 #' @import rlang
 #' @importFrom purrr map map_chr map_dbl map_lgl map_int
 #' @importFrom purrr imap compact map2_chr possibly
+#' @importFrom cli cli_abort cli_text cli_format_method
 #' @importFrom glue glue
 #' @importFrom lifecycle deprecated
 ## usethis namespace: end