+++ R/utils.R
@@ -27,10 +27,6 @@ end_with_slash <- function(x) {
   x
 }
 
-is_testing <- function() {
-  identical(Sys.getenv("TESTTHAT"), "true")
-}
-
 #' Pin Logging
 #'
 #' Log message for diagnosing the `pins` package.
@@ -40,7 +36,7 @@ is_testing <- function() {
 #' @export
 #' @keywords internal
 pin_log <- function(...) {
-  if (getOption("pins.verbose", FALSE) && !is_testing()) {
+  if (getOption("pins.verbose", FALSE)) {
     message(...)
   }
 }
@@ -59,7 +55,7 @@ last <- function(x) x[[length(x)]]
 
 pins_inform <- function(...) {
   opt <- getOption("pins.quiet", NA)
-  if (identical(opt, FALSE) || (identical(opt, NA) && !is_testing())) {
+  if (identical(opt, FALSE) || (identical(opt, NA))) {
     inform(glue(..., .envir = caller_env()))
   }
 }
@@ -97,15 +93,18 @@ envvar_get <- function(name) {
   null_if_na(Sys.getenv(name, NA))
 }
 
-this_not_that <- function(this, that) {
-  abort(glue("Use `{this}` with this board, not `{that}`"))
+this_not_that <- function(this, that, call = caller_env()) {
+  cli_abort(
+    "Use {.fun {this}} with this board, not {.fun {that}}",
+    call = call
+  )
 }
 
-check_board_deparse <- function(board, arg) {
+check_board_deparse <- function(board, arg, call = caller_env()) {
   if (has_name(board, arg)) {
     return(board[[arg]])
   } else {
-    abort(glue("No {arg} found for this board"))
+    cli_abort("No {.arg {arg}} found for this board", call = call)
   }
 }