+++ tests/testthat/setup.R
@@ -1,0 +1,2 @@
+options(pins.verbose = FALSE)
+options(pins.quiet = TRUE)