+++ tests/testthat/test-pin-read-write.R
@@ -8,13 +8,16 @@ test_that("can round trip all types", {
   pin_write(board, df, "df-1", type = "rds")
   expect_equal(pin_read(board, "df-1"), df)
 
-  pin_write(board, df, "df-2", type = "arrow")
+  pin_write(board, df, "df-2", type = "parquet")
   expect_equal(pin_read(board, "df-2"), df)
 
-  pin_write(board, df, "df-3", type = "csv")
+  pin_write(board, df, "df-3", type = "arrow")
+  expect_equal(pin_read(board, "df-2"), df)
+
+  pin_write(board, df, "df-4", type = "csv")
   expect_equal(pin_read(board, "df-3"), df)
 
-  pin_write(board, df, "df-4", type = "qs")
+  pin_write(board, df, "df-5", type = "qs")
   expect_equal(pin_read(board, "df-4"), df)
 
   # List
@@ -51,6 +54,7 @@ test_that("guess_type() works as expected", {
 
 test_that("pin_write() noisily generates name and type", {
   ui_loud()
+  local_mocked_bindings(version_name = function(metadata) "20120304T050607Z-xxxxx")
   expect_snapshot(error = TRUE, {
     b <- board_temp()
     pin_write(b, mtcars)
@@ -68,6 +72,10 @@ test_that("user can supply metadata", {
 })
 
 test_that("can request specific hash", {
+  local_mocked_bindings(
+    version_name = function(metadata) "20120304T050607Z-xxxxx",
+    write_rds = write_rds_test
+    )
   ui_loud()
   expect_snapshot(error = TRUE, {
     b <- board_temp()
@@ -83,3 +91,4 @@ test_that("informative error for legacy boards", {
     board %>% pin_read("x")
   })
 })
+