+++ tests/testthat/test-pin_versions.R
@@ -47,8 +47,16 @@ test_that("informative error for writing with same version", {
 # versions pruning --------------------------------------------------------
 
 test_that("can prune old versions", {
-  board <- board_temp(versioned = TRUE)
+  skip_if_not_installed("mockery")
 
+  board <- board_temp(versioned = TRUE)
+  mock_version_name <- mockery::mock(
+    "20130104T050607Z-xxxxx",
+    "20130204T050607Z-yyyyy",
+    "20130304T050607Z-zzzzz",
+    "20130404T050607Z-aaaaa"
+  )
+  local_mocked_bindings(version_name = mock_version_name)
   pin_write(board, 1, "x")
   pin_write(board, 2, "x")
   pin_write(board, 3, "x")